refactor: centralize nodeOutputs, DBOutput consolidate, split substrate persistence
Made-with: Cursor
