+++ apps/agate-ui/src/pages/RunGraph.tsx
@@ -22,7 +22,7 @@ import NodePalette from '@/components/NodePalette'
 import RunPanel from '@/components/RunPanel'
 import ConfirmDialog from '@/components/ConfirmDialog'
 import { getGraph, createRun, getRun, updateGraph, deleteGraph, type Graph, type Run } from '@/lib/api'
-import type { NodeOutputLookupSpec } from '@/lib/nodeOutputs'
+import { nodeOutputLookupFromReactFlow } from '@/lib/nodeOutputs'
 import { ArrowLeft, Save, Edit, Loader2, Play, Trash2 } from 'lucide-react'
 
 export default function RunGraph() {
@@ -62,11 +62,8 @@ export default function RunGraph() {
     onCancel?: () => void
   } | null>(null)
   
-  const nodeOutputLookupSpec: NodeOutputLookupSpec = useMemo(
-    () => ({
-      nodes: nodes.map((n) => ({ id: n.id!, type: n.type! })),
-      edges: edges.map((e) => ({ source: e.source, target: e.target })),
-    }),
+  const nodeOutputLookupSpec = useMemo(
+    () => nodeOutputLookupFromReactFlow(nodes, edges),
     [nodes, edges],
   )
