+++ apps/worker/src/worker/nodes/db_output.py
@@ -5,24 +5,10 @@ from __future__ import annotations
 import os
 from typing import Any
 
-from agate_runtime.output_node import (
-    OutputConsolidator,
-    OutputParams,
-    expand_upstream_merge_for_output_consolidator,
-)
+from agate_runtime.output_node import consolidated_body_from_dboutput
 from sqlmodel import Session
 
 from worker.substrate_persistence import persist_from_consolidated
-
-
-def _merge_namespaced_upstream_inputs(inputs: dict[str, Any]) -> dict[str, Any]:
-    """Shallow-merge upstream node outputs keyed by upstream node id."""
-
-    merged: dict[str, Any] = {}
-    for _upstream_id, payload in inputs.items():
-        if isinstance(payload, dict):
-            merged.update(payload)
-    return merged
 
 
 def run_db_output(params: dict[str, Any], inputs: dict[str, Any]) -> dict[str, Any]:
@@ -36,19 +22,13 @@ def run_db_output(params: dict[str, Any], inputs: dict[str, Any]) -> dict[str, A
         )
     project_id = int(project_id_raw)
 
-    merged = _merge_namespaced_upstream_inputs(inputs)
-    merged = expand_upstream_merge_for_output_consolidator(merged)
-    cons = OutputConsolidator()
-    p = OutputParams.model_validate(params)
-    body = cons.run(merged, p.model_dump())
+    body = consolidated_body_from_dboutput(params, inputs)
 
-    engine_url = os.getenv("BACKFIELD_DATABASE_URL") or os.getenv("DATABASE_URL")
-    if not engine_url:
-        raise RuntimeError("Missing BACKFIELD_DATABASE_URL / DATABASE_URL for DBOutput persistence")
-
+    from backfield_db.session import get_database_url
     from sqlalchemy.engine import make_url
     from sqlmodel import create_engine
 
+    engine_url = get_database_url()
     connect_args: dict[str, Any] = {}
     try:
         url = make_url(engine_url)
