+++ packages/backfield-core/src/backfield_core/nodes/db_output.py
@@ -5,23 +5,11 @@ from __future__ import annotations
 import os
 from typing import Any
 
-from agate_runtime.output_node import (
-    OutputConsolidator,
-    OutputParams,
-    expand_upstream_merge_for_output_consolidator,
-)
+from agate_runtime.output_node import consolidated_body_from_dboutput
 
 
 def run_db_output(params: dict[str, Any], inputs: dict[str, Any]) -> dict[str, Any]:
-    merged: dict[str, Any] = {}
-    for _upstream_id, payload in inputs.items():
-        if isinstance(payload, dict):
-            merged.update(payload)
-    merged = expand_upstream_merge_for_output_consolidator(merged)
-
-    cons = OutputConsolidator()
-    p = OutputParams.model_validate(params)
-    body = cons.run(merged, p.model_dump())
+    body = consolidated_body_from_dboutput(params, inputs)
 
     # The Celery worker registers a different `DBOutput` runner that writes Postgres.
     # In `backfield-core` unit tests (and other non-worker hosts), treat this as a no-op that
