+++ packages/backfield-ui/package.json
@@ -6,7 +6,8 @@
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./nodeOutputs": "./src/nodeOutputs.ts"
   },
   "peerDependencies": {
     "react": "^18.2.0",
