+++ packages/backfield-ui/src/index.ts
@@ -1,2 +1,11 @@
 export { UserAccountMenu, type UserAccountMenuProps } from "./UserAccountMenu"
 export { cn } from "./cn"
+export {
+  NODE_OUTPUT_KEY_INDEX,
+  buildNodeIdToPublicOutputKey,
+  getNodeOutputById,
+  getNodeOutputKeyMap,
+  nodeOutputLookupFromGraphSpec,
+  nodeOutputLookupFromReactFlow,
+  type NodeOutputLookupSpec,
+} from "./nodeOutputs"
